requirements.txt
psutil==5.9.1
discord.py==2.0.0